refactor：move starlight bonus related classes and constants to a separate file to reduce clutter in team_setting_card.py
